Remove unnecessary deffinition of innerButtonBlock
The example is style: 'inline', which is the actual widget embed, where the Button is not displayed. So, defining the innerBlocks here is unnecessary. We can just define it within the edit.js where it is used instead.
